+++ sdk/c4-client.ts
@@ -2,8 +2,8 @@
 // Generated from /openapi.json via src/openapi-sdk-gen.js.
 // Do not edit by hand — re-run `c4 openapi --sdk` to refresh.
 
-// Spec version: 1.10.59
-// Generated at: 2026-05-02T02:58:06.101Z
+// Spec version: 1.10.60
+// Generated at: 2026-05-02T03:01:37.896Z
 
 export interface postAuthLoginBody {
   user: string; /** Username */